# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Почтовый сервис",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
        "/mailer/get": {
            "post": {
                "description": "Получение сообщений из почты пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mailer"
                ],
                "summary": "GetMail",
                "operationId": "mailer-get-mail",
                "parameters": [
                    {
                        "description": "Настройки",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/message.OutputModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "messages",
                        "schema": {
                            "$ref": "#/definitions/message.MessagesModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/mailer/send": {
            "post": {
                "description": "Отправка сообщения пользователю",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mailer"
                ],
                "summary": "SendMail",
                "operationId": "mailer-send-mail",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Файлы",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Адрес получателя сообщения",
                        "name": "receiver",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Сообщение для отправки",
                        "name": "message",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "file.FileModel": {
            "type": "object",
            "properties": {
                "filename": {
                    "description": "Имя файлы",
                    "type": "string"
                },
                "filepath": {
                    "description": "Полный путь к файлу",
                    "type": "string"
                }
            }
        },
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "message.MessageOutputModel": {
            "type": "object",
            "required": [
                "message",
                "sender",
                "subject"
            ],
            "properties": {
                "files": {
                    "description": "Прикреплённые файлы",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/file.FileModel"
                    }
                },
                "message": {
                    "description": "Тело сообщения",
                    "type": "string"
                },
                "sender": {
                    "description": "Отправитель сообщения",
                    "type": "string"
                },
                "subject": {
                    "description": "Тема сообщения",
                    "type": "string"
                }
            }
        },
        "message.MessagesModel": {
            "type": "object",
            "properties": {
                "messages": {
                    "description": "Множество сообщений",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/message.MessageOutputModel"
                    }
                }
            }
        },
        "message.OutputModel": {
            "type": "object",
            "required": [
                "count"
            ],
            "properties": {
                "count": {
                    "description": "Количество считываемых сообщений",
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}