# pkg/model/file/file.go
package file

/**
 * Структура, описывающая содержимое информации о файле
 */
type FileModel struct {
	Filename string // Имя файлы
	Filepath string // Полный путь к файлу
}